[PATCH] misc fixes
- mpparse printk should be in hex (john stultz <johnstul@us.ibm.com>)

- fiddle with RCU copyright messages (Dipankar Sarma <dipankar@in.ibm.com>)

- use print_dev_t() for sysfs dev file in videodev.c (Gerd Knorr
  <kraxel@bytesex.org>)

- comx-hw-munich.c 64-bit warning fix (Vinay K Nallamothu
  <vinay.nallamothu@gsecone.com>)

- random.c return val fix

+++ drivers/net/wan/comx-hw-munich.c
@@ -1849,7 +1849,7 @@ static int MUNICH_open(struct net_device *dev)
     if (board->isx21)
     {
 	init_timer(&board->modemline_timer);
-	board->modemline_timer.data = (unsigned int)board;
+	board->modemline_timer.data = (unsigned long)board;
 	board->modemline_timer.function = pcicom_modemline;
 	board->modemline_timer.expires = jiffies + HZ;
 	add_timer((struct timer_list *)&board->modemline_timer);
